Enhancing file access

+++ backend/utils/db.py
@@ -1,9 +1,8 @@
 from copy import copy
 import shutil
-from models.enums import LockAction
 from utils.helpers import arr_remove_common, branch_path, snake_case
 from datetime import datetime
-from models.enums import ContentType, ResourceType
+from models.enums import ContentType, ResourceType, LockAction
 from utils.internal_error_code import InternalErrorCode
 from utils.middleware import get_request_data
 from utils.redis_services import RedisServices
@@ -11,13 +10,13 @@ from utils.settings import settings
 import models.core as core
 from typing import TypeVar, Type, Any
 import models.api as api
-import os
 import json
 from pathlib import Path
 from fastapi import status
 import aiofiles
 from utils.regex import FILE_PATTERN, FOLDER_PATTERN
 from shutil import copy2 as copy_file
+import os
 
 MetaChild = TypeVar("MetaChild", bound=core.Meta)
 
@@ -236,12 +235,13 @@ async def load(
         )
 
     path /= filename
-    async with aiofiles.open(path, "r") as file:
-        content = await file.read()
-        try:
+    content = ""
+    try: 
+        async with aiofiles.open(path, "r") as file:
+            content = await file.read()
             return class_type.model_validate_json(content)
-        except Exception as e:
-            raise Exception(f"Error Invalid Entry At: {path}. Error {e}")
+    except Exception as e:
+        raise Exception(f"Error Invalid Entry At: {path}. Error {e} {content=}")
 
 
 def load_resource_payload(
@@ -259,11 +259,15 @@ def load_resource_payload(
     path /= filename
     if not path.is_file():
         return {}
-        # raise api.Exception(
-        #     status_code=status.HTTP_404_NOT_FOUND,
-        #     error=api.Error(type="db", code=12, message="Request object is not available"),
-        # )
-    return json.loads(path.read_bytes())
+    try: 
+        bytes = path.read_bytes()
+        data = json.loads(bytes)
+        return data
+    except Exception as _:
+        raise api.Exception(
+          status_code=status.HTTP_404_NOT_FOUND,
+          error=api.Error(type="db", code=12, message=f"Request object is not available {path}"),
+        )
 
 
 async def save(
@@ -283,8 +287,10 @@ async def save(
         os.makedirs(path)
 
     meta_json = meta.model_dump_json(exclude_none=True)
-    async with aiofiles.open(path / filename, "w") as file:
-        await file.write(meta_json)
+    with open(path / filename, "w") as file:
+        file.write(meta_json)
+        file.flush()
+        os.fsync(file)
 
 
 async def create(
@@ -304,8 +310,10 @@ async def create(
     if not path.is_dir():
         os.makedirs(path)
 
-    async with aiofiles.open(path / filename, "w") as file:
-        await file.write(meta.model_dump_json(exclude_none=True))
+    with open(path / filename, "w") as file:
+        file.write(meta.model_dump_json(exclude_none=True))
+        file.flush()
+        os.fsync(file)
 
 
 async def save_payload(
@@ -326,8 +334,10 @@ async def save_payload(
         )
 
     content = await attachment.read()
-    async with aiofiles.open(payload_file_path / payload_filename, "wb") as file:
-        await file.write(content)
+    with open(payload_file_path / payload_filename, "wb") as file:
+        file.write(content)
+        file.flush()
+        os.fsync(file)
 
 
 async def save_payload_from_json(
@@ -363,8 +373,10 @@ async def save_payload_from_json(
         )
 
     payload_json = json.dumps(payload_data)
-    async with aiofiles.open(payload_file_path / payload_filename, "w") as file:
-        await file.write(payload_json)
+    with open(payload_file_path / payload_filename, "w") as file:
+        file.write(payload_json)
+        file.flush()
+        os.fsync(file)
 
 
 async def update(
@@ -377,6 +389,7 @@ async def update(
     branch_name: str | None,
     user_shortname: str,
     schema_shortname: str | None = None,
+    retrieve_lock_status: bool | None = False,
 ) -> dict:
     """Update the entry, store the difference and return it"""
     path, filename = metapath(
@@ -393,38 +406,41 @@ async def update(
             error=api.Error(type="update", code=InternalErrorCode.OBJECT_NOT_FOUND,
                             message="Request object is not available"),
         )
-    async with RedisServices() as redis_services:
-        if await redis_services.is_entry_locked(
-            space_name, branch_name, subpath, meta.shortname, user_shortname
-        ):
-            raise api.Exception(
-                status_code=status.HTTP_403_FORBIDDEN,
-                error=api.Error(
-                    type="update", code=InternalErrorCode.LOCKED_ENTRY, message="This entry is locked"),
-            )
-        elif await redis_services.get_lock_doc(
-            space_name, branch_name, subpath, meta.shortname
-        ):
-            # if the current can release the lock that means he is the right user
-            await redis_services.delete_lock_doc(
+    if retrieve_lock_status:
+        async with RedisServices() as redis_services:
+            if await redis_services.is_entry_locked(
+                space_name, branch_name, subpath, meta.shortname, user_shortname
+            ):
+                raise api.Exception(
+                    status_code=status.HTTP_403_FORBIDDEN,
+                    error=api.Error(
+                        type="update", code=InternalErrorCode.LOCKED_ENTRY, message="This entry is locked"),
+                )
+            elif await redis_services.get_lock_doc(
                 space_name, branch_name, subpath, meta.shortname
-            )
-            await store_entry_diff(
-                space_name,
-                branch_name,
-                "/" + subpath,
-                meta.shortname,
-                user_shortname,
-                {},
-                {"lock_type": LockAction.unlock},
-                ["lock_type"],
-                core.Content,
-            )
+            ):
+                # if the current can release the lock that means he is the right user
+                await redis_services.delete_lock_doc(
+                    space_name, branch_name, subpath, meta.shortname
+                )
+                await store_entry_diff(
+                    space_name,
+                    branch_name,
+                    "/" + subpath,
+                    meta.shortname,
+                    user_shortname,
+                    {},
+                    {"lock_type": LockAction.unlock},
+                    ["lock_type"],
+                    core.Content,
+                )
 
     meta.updated_at = datetime.now()
     meta_json = meta.model_dump_json(exclude_none=True)
-    async with aiofiles.open(path / filename, "w") as file:
-        await file.write(meta_json)
+    with open(path / filename, "w") as file:
+        file.write(meta_json)
+        file.flush()
+        os.fsync(file)
 
     history_diff = await store_entry_diff(
         space_name,
@@ -610,8 +626,10 @@ async def move(
 
     if meta_updated:
         meta_json = meta.model_dump_json(exclude_none=True)
-        async with aiofiles.open(dest_path / dest_filename, "w") as opened_file:
-            await opened_file.write(meta_json)
+        with open(dest_path / dest_filename, "w") as opened_file:
+            opened_file.write(meta_json)
+            opened_file.flush()
+            os.fsync(opened_file)
 
     # Delete Src path if empty
     if src_path.parent.is_dir():
@@ -688,6 +706,7 @@ async def delete(
     branch_name: str | None,
     user_shortname: str,
     schema_shortname: str | None = None,
+    retrieve_lock_status: bool | None = False,
 ):
     """Delete the file that match the criteria given, remove folder if empty
 
@@ -718,20 +737,21 @@ async def delete(
             error=api.Error(
                 type="delete", code=InternalErrorCode.OBJECT_NOT_FOUND, message="Request object is not available"),
         )
-    async with RedisServices() as redis_services:
-        if await redis_services.is_entry_locked(
-            space_name, branch_name, subpath, meta.shortname, user_shortname
-        ):
-            raise api.Exception(
-                status_code=status.HTTP_403_FORBIDDEN,
-                error=api.Error(
-                    type="delete", code=InternalErrorCode.LOCKED_ENTRY, message="This entry is locked"),
-            )
-        else:
-            # if the current can release the lock that means he is the right user
-            await redis_services.delete_lock_doc(
-                space_name, branch_name, subpath, meta.shortname
-            )
+    if retrieve_lock_status:
+        async with RedisServices() as redis_services:
+            if await redis_services.is_entry_locked(
+                    space_name, branch_name, subpath, meta.shortname, user_shortname
+            ):
+                raise api.Exception(
+                    status_code=status.HTTP_403_FORBIDDEN,
+                    error=api.Error(
+                        type="delete", code=InternalErrorCode.LOCKED_ENTRY, message="This entry is locked"),
+                )
+            else:
+                # if the current can release the lock that means he is the right user
+                await redis_services.delete_lock_doc(
+                    space_name, branch_name, subpath, meta.shortname
+                )
 
     pathname = path / filename
     if pathname.is_file():
